cookbooks/python/browser-use/requirements.txt
evolve-sdk
python-dotenv
pydantic
